+++ packages/charity-frontend/package.json
@@ -9,9 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@types/node": "20.2.5",
-    "@types/react": "18.2.7",
-    "@types/react-dom": "18.2.4",
+    "@ant-design/icons": "^5.1.4",
     "antd": "^5.5.1",
     "autoprefixer": "10.4.14",
     "eslint": "8.41.0",
@@ -21,7 +19,13 @@
     "postcss": "8.4.23",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "tailwindcss": "3.3.2",
+    "tailwindcss": "3.3.2"
+  },
+  "devDependencies": {
+    "sass": "^1.62.1",
+    "@types/node": "20.2.5",
+    "@types/react": "18.2.7",
+    "@types/react-dom": "18.2.4",
     "typescript": "5.0.4"
   }
 }
